--- src/actions.js
@@ -1,20 +1,0 @@
-export {
-  setRegions,
-  setCategories,
-  setRestaurants,
-  setRestaurant,
-  selectRegion,
-  selectCategory,
-  setAccessToken,
-  setReviews,
-  logout,
-  clearReviewFields,
-  changeLoginField,
-  changeReviewField,
-  loadInitialData,
-  loadRestaurants,
-  loadRestaurant,
-  requestLogin,
-  loadReview,
-  sendReview,
-} from './slice';
