+++ src/containers/RestaurantContainer.jsx
@@ -10,7 +10,7 @@ import {
   loadRestaurant,
   changeReviewField,
   sendReview,
-} from '../actions';
+} from '../slice';
 
 import { get } from '../utils';
